requirements.txt
Django>=3.0,<=3.1
Keras==2.4.3
numpy
Pillow
pip==21.0.1
PyYAML
scipy
tensorflow
setuptools==56.0.0